Move woocommerce/product-tab to blocks folder
