+++ packages/js/product-editor/src/blocks/tab/block.json
