+++ packages/js/product-editor/src/style.scss
@@ -5,7 +5,7 @@
 /* Structural */
 
 @import 'blocks/section/editor.scss';
-@import 'components/tab/editor.scss';
+@import 'blocks/tab/editor.scss';
 @import 'components/tabs/style.scss';
 @import 'components/product-section-layout/style.scss';
 @import 'components/header/style.scss';
